+++ SacNew/Models/POC.cs
@@ -6,7 +6,7 @@
         public string NumeroPOC { get; set; }
         public int IdPosta { get; set; }
         public int IdNomina { get; set; }
-        public decimal Odometro { get; set; }
+        public double Odometro { get; set; }
         public string Comentario { get; set; }
         public DateTime FechaCreacion { get; set; }
         public int IdUsuario { get; set; }
